update source data, tests, sync fixtures
