+++ refiner/tests/fixtures/eicr_v1_1/mon_mothma_covid_influenza_eICR.xml
@@ -358,13 +358,6 @@
               <code code="99213" codeSystem="2.16.840.1.113883.6.12" codeSystemName="CPT-4"
                 displayName="Office outpatient visit 15 minutes" />
               <effectiveTime value="20250216" />
-              <participant typeCode="LOC">
-                <participantRole classCode="SDLOC">
-                  <code code="1160-1" codeSystem="2.16.840.1.113883.1.11.20275"
-                    codeSystemName="HealthcareServiceLocation"
-                    displayName="Grand Republic Medical Facility Clinic" />
-                </participantRole>
-              </participant>
               <performer>
                 <assignedEntity>
                   <id extension="1234567890" root="2.16.840.1.113883.4.6" />
@@ -377,6 +370,14 @@
                   </assignedPerson>
                 </assignedEntity>
               </performer>
+              <participant typeCode="LOC">
+                <participantRole classCode="SDLOC">
+                  <code code="1160-1" codeSystem="2.16.840.1.113883.1.11.20275"
+                    codeSystemName="HealthcareServiceLocation"
+                    displayName="Grand Republic Medical Facility Clinic" />
+                </participantRole>
+              </participant>
+
             </encounter>
           </entry>
           <!-- Follow-up Visit Entry -->
@@ -388,13 +389,6 @@
               <code code="99213" codeSystem="2.16.840.1.113883.6.12" codeSystemName="CPT-4"
                 displayName="Office outpatient visit - followup" />
               <effectiveTime value="20250217" />
-              <participant typeCode="LOC">
-                <participantRole classCode="SDLOC">
-                  <code code="1160-1" codeSystem="2.16.840.1.113883.1.11.20275"
-                    codeSystemName="HealthcareServiceLocation"
-                    displayName="Grand Republic Medical Facility Clinic" />
-                </participantRole>
-              </participant>
               <performer>
                 <assignedEntity>
                   <id extension="1234567890" root="2.16.840.1.113883.4.6" />
@@ -407,6 +401,14 @@
                   </assignedPerson>
                 </assignedEntity>
               </performer>
+              <participant typeCode="LOC">
+                <participantRole classCode="SDLOC">
+                  <code code="1160-1" codeSystem="2.16.840.1.113883.1.11.20275"
+                    codeSystemName="HealthcareServiceLocation"
+                    displayName="Grand Republic Medical Facility Clinic" />
+                </participantRole>
+              </participant>
+
             </encounter>
           </entry>
           <!-- Results Visit with Diagnoses -->
@@ -418,13 +420,6 @@
               <code code="99212" codeSystem="2.16.840.1.113883.6.12" codeSystemName="CPT-4"
                 displayName="Office outpatient visit - results review" />
               <effectiveTime value="20250221" />
-              <participant typeCode="LOC">
-                <participantRole classCode="SDLOC">
-                  <code code="1160-1" codeSystem="2.16.840.1.113883.1.11.20275"
-                    codeSystemName="HealthcareServiceLocation"
-                    displayName="Grand Republic Medical Facility Clinic" />
-                </participantRole>
-              </participant>
               <performer>
                 <assignedEntity>
                   <id extension="1234567890" root="2.16.840.1.113883.4.6" />
@@ -437,6 +432,14 @@
                   </assignedPerson>
                 </assignedEntity>
               </performer>
+              <participant typeCode="LOC">
+                <participantRole classCode="SDLOC">
+                  <code code="1160-1" codeSystem="2.16.840.1.113883.1.11.20275"
+                    codeSystemName="HealthcareServiceLocation"
+                    displayName="Grand Republic Medical Facility Clinic" />
+                </participantRole>
+              </participant>
+
               <!-- Link to COVID-19 Diagnosis -->
               <entryRelationship typeCode="COMP">
                 <act classCode="ACT" moodCode="EVN">
